refactor: move EventIpcManagerEnrichingWrapper to daemon-common, wire for Pollerd
Move EventIpcManagerEnrichingWrapper from daemon-boot-eventtranslator to
daemon-common so any event-producing daemon can use it. This wrapper
intercepts sendNow() calls and enriches events with alarm-data, severity,
and logmsg from EventConfEnrichmentService before they reach Kafka.

Pollerd now wraps its EventIpcManager with the enriching wrapper, matching
the same pattern EventTranslator already uses. This replaces the previous
approach of wiring EventConfDao into KafkaEventForwarder.

Also removes the hardcoded addAlarmData() from DefaultPollContext —
eventconf enrichment handles all UEIs generically.
