+++ core/daemon-common/src/main/java/org/opennms/core/daemon/common/EventConfEnrichmentService.java
@@ -73,14 +73,6 @@ public class EventConfEnrichmentService {
     }
 
     /**
-     * Returns the underlying EventConfDao for use by other components (e.g.,
-     * KafkaEventForwarder) that need to look up event configurations.
-     */
-    public DefaultEventConfDao getEventConfDao() {
-        return eventConfDao;
-    }
-
-    /**
      * Enriches an event by looking up the matching event conf entry and applying
      * alarm-data, severity, and logmsg when the event is missing them.
      *
