+++ core/daemon-common/src/main/java/org/opennms/core/daemon/common/KafkaEventTransportConfiguration.java
@@ -67,14 +67,9 @@ public class KafkaEventTransportConfiguration {
     private long pollTimeoutMs;
 
     @Bean
-    public KafkaEventForwarder kafkaEventForwarder(
-            @org.springframework.beans.factory.annotation.Autowired(required = false)
-            EventConfEnrichmentService eventConfEnrichmentService) {
+    public KafkaEventForwarder kafkaEventForwarder() {
         KafkaEventForwarder forwarder = KafkaEventForwarderFactory.create(bootstrapServers, eventTopic);
         forwarder.setIpcTopicName(ipcTopic);
-        if (eventConfEnrichmentService != null) {
-            forwarder.setEventConfDao(eventConfEnrichmentService.getEventConfDao());
-        }
         return forwarder;
     }
